(refactor) : 각 라운드의 결과들과 우승자를 GameManagerResponseDto에 담아서 controller에 반환하도록 구현
