+++ src/main/java/racingcar/controller/GameManagerController.java
@@ -1,8 +1,8 @@
 package racingcar.controller;
 
 import racingcar.domain.gamemanager.GameManager;
-import racingcar.domain.result.Results;
-import racingcar.domain.gamemanager.dto.GameManagerRequestDto;
+import racingcar.controller.dto.GameManagerResponseDto;
+import racingcar.controller.dto.GameManagerRequestDto;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
 
@@ -10,7 +10,7 @@ public class GameManagerController {
     public static void main(String[] args) {
         GameManagerRequestDto requestDto = InputView.inputFromUser();
         GameManager gameManager = new GameManager();
-        Results results = gameManager.playGame(requestDto);
-        OutputView.printResults(results);
+        GameManagerResponseDto responseDto = gameManager.playGame(requestDto);
+        OutputView.printResults(responseDto);
     }
 }
